+++ blocklets/core/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.4.31 (2025-9-17)
 
-- fix: update credential
+- fix: improve credential validation and status checking
 
 ## 0.4.30 (2025-9-17)